+++ upnp/src/inc/uri.h
@@ -289,7 +289,7 @@ int parse_hostport(
 	/*! [in] Sets a maximum limit. */
 	size_t max,
     /*! [in] The default port if the port is not specified. */
-    int defaultPort,
+    unsigned short int defaultPort,
 	/*! [out] Output parameter where the host and port are represented as
 	 * an internet address. */
 	hostport_type *out);
